refactor-app: 디렉토리 계층 구조 적용
