+++ application/lib/common/router/router_provider.dart
@@ -3,14 +3,15 @@ import 'package:application_new/common/session/session_provider.dart';
 import 'package:application_new/feature/authentication/page/login_page.dart';
 import 'package:application_new/feature/home/home_page.dart';
 import 'package:application_new/feature/travel_create/page/travel_create_page.dart';
-import 'package:application_new/feature/travel_plan/city_place_pois/page/city_place_pois_page.dart';
-import 'package:application_new/feature/travel_plan/city_travels/page/city_travels_page.dart';
+import 'package:application_new/feature/travel_plan/page/travel_plan_recommend/page/city_travels/page/city_travels_page.dart';
 import 'package:application_new/feature/travel_read/page/travel_read_page.dart';
 import 'package:application_new/feature/travel_list/page/travel_list_page.dart';
 import 'package:application_new/feature/travel_plan/page/travel_plan_page.dart';
 import 'package:go_router/go_router.dart';
 
 import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+import '../../feature/travel_plan/page/travel_plan_recommend/page/city_place_pois/page/city_place_pois_page.dart';
 
 part 'router_provider.g.dart';
