+++ application/lib/feature/travel_plan/page/travel_plan_page.dart
@@ -6,9 +6,8 @@ import 'package:application_new/feature/travel_plan/page/travel_plan_bookmark_pa
 import 'package:application_new/feature/travel_plan/page/travel_plan_home_page.dart';
 import 'package:application_new/feature/travel_plan/page/travel_plan_manage_page.dart';
 import 'package:application_new/feature/travel_plan/provider/travel_plan_state.dart';
-import 'package:application_new/feature/travel_plan/travel_plan_recommend/page/travel_plan_recommend_page.dart';
+import 'package:application_new/feature/travel_plan/page/travel_plan_recommend/page/travel_plan_recommend_page.dart';
 import 'package:application_new/feature/travel_plan/provider/travel_plan_provider.dart';
-import 'package:application_new/domain/travel/travel_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
