+++ application/lib/feature/travel_plan/page/travel_plan_recommend/component/place_metric_item.dart
@@ -1,11 +1,12 @@
 import 'package:application_new/common/util/string_extension.dart';
 import 'package:application_new/common/util/translation_util.dart';
-import 'package:application_new/feature/travel_plan/city_place_pois/provider/city_place_pois_state.dart';
 import 'package:application_new/shared/component/outlined_icon_button.dart';
 import 'package:application_new/shared/component/small_chip.dart';
 import 'package:extended_wrap/extended_wrap.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import '../page/city_place_pois/provider/city_place_pois_state.dart';
 
 class PlaceMetricItem extends ConsumerWidget {
