+++ application/lib/feature/travel_plan/page/travel_plan_recommend/component/travel_item.dart
