+++ application/lib/feature/travel_plan/page/travel_plan_recommend/model/place_recommend_model.freezed.dart
