+++ application/lib/feature/travel_plan/page/travel_plan_recommend/page/city_place_pois/component/place_metric_card_item.dart
@@ -1,9 +1,10 @@
 
 import 'package:application_new/common/util/string_extension.dart';
 import 'package:application_new/common/util/translation_util.dart';
-import 'package:application_new/feature/travel_plan/city_place_pois/provider/city_place_pois_state.dart';
 import 'package:application_new/shared/component/small_chip.dart';
 import 'package:flutter/material.dart';
+
+import '../provider/city_place_pois_state.dart';
 
 class PlaceMetricCardItem extends StatelessWidget {
