+++ application/lib/feature/travel_plan/page/travel_plan_recommend/page/city_place_pois/provider/city_place_map_state.freezed.dart
