+++ application/lib/feature/travel_plan/page/travel_plan_recommend/page/city_travels/model/paged_travels_model.freezed.dart
