+++ application/lib/feature/travel_plan/page/travel_plan_recommend/page/city_travels/provider/city_travels_state.dart
@@ -1,6 +1,6 @@
 import 'package:application_new/domain/geography/geography_model.dart';
-import 'package:application_new/feature/travel_plan/city_travels/model/paged_travels_model.dart';
 import 'package:application_new/domain/travel/travel_model.dart';
+import 'package:application_new/feature/travel_plan/page/travel_plan_recommend/page/city_travels/model/paged_travels_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'city_travels_state.freezed.dart';
