+++ application/lib/feature/travel_plan/page/travel_plan_recommend/page/city_travels/provider/city_travels_state.freezed.dart
