+++ application/lib/feature/travel_plan/page/travel_plan_recommend/page/sliver_city_poi_preview.dart
@@ -1,12 +1,12 @@
-import 'package:application_new/common/loading/loading_page.dart';
 import 'package:application_new/common/util/translation_util.dart';
-import 'package:application_new/feature/travel_plan/city_place_pois/provider/city_place_pois_provider.dart';
-import 'package:application_new/feature/travel_plan/city_place_pois/provider/city_place_pois_state.dart';
-import 'package:application_new/feature/travel_plan/travel_plan_recommend/component/place_metric_item.dart';
+import 'package:application_new/feature/travel_plan/page/travel_plan_recommend/component/place_metric_item.dart';
 import 'package:application_new/shared/util/snap_scroll_physics.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
+
+import 'city_place_pois/provider/city_place_pois_provider.dart';
+import 'city_place_pois/provider/city_place_pois_state.dart';
 
 
 class SliverCityPoiPreview extends ConsumerWidget {
