+++ application/lib/feature/travel_plan/page/travel_plan_recommend/page/sliver_city_travel_preview.dart
@@ -1,11 +1,9 @@
 import 'dart:math';
 
 import 'package:application_new/common/util/translation_util.dart';
-import 'package:application_new/domain/geography/geography_model.dart';
-import 'package:application_new/feature/travel_plan/city_travels/provider/city_travels_provider.dart';
-import 'package:application_new/feature/travel_plan/city_travels/provider/city_travels_state.dart';
-import 'package:application_new/feature/travel_plan/travel_plan_recommend/component/travel_item.dart';
-import 'package:application_new/domain/travel/travel_model.dart';
+import 'package:application_new/feature/travel_plan/page/travel_plan_recommend/component/travel_item.dart';
+import 'package:application_new/feature/travel_plan/page/travel_plan_recommend/page/city_travels/provider/city_travels_provider.dart';
+import 'package:application_new/feature/travel_plan/page/travel_plan_recommend/page/city_travels/provider/city_travels_state.dart';
 import 'package:application_new/shared/util/constants.dart';
 import 'package:application_new/shared/util/snap_scroll_physics.dart';
 import 'package:flutter/material.dart';
