+++ application/lib/feature/travel_plan/page/travel_plan_recommend/provider/travel_plan_recommend_provider.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:application_new/common/http/http_service_provider.dart';
 import 'package:application_new/common/util/iterable_util.dart';
 import 'package:application_new/feature/travel_plan/provider/travel_plan_provider.dart';
-import 'package:application_new/feature/travel_plan/travel_plan_recommend/model/place_recommend_model.dart';
+import 'package:application_new/feature/travel_plan/page/travel_plan_recommend/model/place_recommend_model.dart';
 import 'package:application_new/domain/place/place_model.dart';
 import 'package:application_new/domain/travel/travel_model.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
