+++ application/lib/feature/travel_plan/page/travel_plan_recommend/provider/travel_plan_recommend_state.dart
@@ -1,5 +1,5 @@
 import 'package:application_new/domain/geography/geography_model.dart';
-import 'package:application_new/feature/travel_plan/travel_plan_recommend/model/place_recommend_model.dart';
+import 'package:application_new/feature/travel_plan/page/travel_plan_recommend/model/place_recommend_model.dart';
 import 'package:application_new/domain/place/place_model.dart';
 import 'package:application_new/domain/travel/travel_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
