+++ application/lib/feature/travel_plan/page/travel_plan_recommend/provider/travel_plan_recommend_state.freezed.dart
